Refactor visualize wizard. - Migrate from absolute imports to relative imports. - Reorganize files for clearer hierarchy. - Represent one concern per file. - Aggregate route definitions in one place. - Clarify entry points into code, with clear dependencies.
