+++ src/core_plugins/kibana/public/visualize/wizard/step_2.html
@@ -1,22 +1,24 @@
-<bread-crumbs></bread-crumbs>
-<div class="wizard">
-  <div class="wizard-column wizard-column--small">
-    <h3 class="wizard-sub-title">From a New Search, Select Index</h3>
-    <paginated-selectable-list
-      per-page="20"
-      list="indexPattern.list"
-      user-make-url="makeUrl"
-      class="wizard-row"
-    ></paginated-selectable-list>
-  </div>
-  <div class="wizard-column wizard-column--large">
-    <h3 class="wizard-sub-title">Or, From a Saved Search</h3>
-    <!-- Saved searches -->
-    <saved-object-finder
-        title="Saved Searches"
-        type="searches"
+<div class="container-fluid vis-wizard">
+  <bread-crumbs></bread-crumbs>
+  <div class="wizard">
+    <div class="wizard-column wizard-column--small">
+      <h3 class="wizard-sub-title">From a New Search, Select Index</h3>
+      <paginated-selectable-list
+        per-page="20"
+        list="indexPattern.list"
+        user-make-url="makeUrl"
         class="wizard-row"
-        make-url="step2WithSearchUrl"
-      ></saved-object-finder>
+      ></paginated-selectable-list>
+    </div>
+    <div class="wizard-column wizard-column--large">
+      <h3 class="wizard-sub-title">Or, From a Saved Search</h3>
+      <!-- Saved searches -->
+      <saved-object-finder
+          title="Saved Searches"
+          type="searches"
+          class="wizard-row"
+          make-url="step2WithSearchUrl"
+        ></saved-object-finder>
+    </div>
   </div>
 </div>
